# web/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/templates": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": [
                            "templates.view"
                        ]
                    }
                ],
                "description": "Gets all repos that are available to pull template from",
                "tags": [
                    "Templates"
                ],
                "summary": "Get all repos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TemplateRepo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OAuth2Application": [
                            "templates.repo.create"
                        ]
                    }
                ],
                "description": "Adds a new repo to the service",
                "tags": [
                    "Templates"
                ],
                "summary": "Add repo",
                "parameters": [
                    {
                        "description": "Repo information",
                        "name": "repo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TemplateRepo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TemplateRepo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/templates/local/{name}": {
            "put": {
                "security": [
                    {
                        "OAuth2Application": [
                            "templates.local.edit"
                        ]
                    }
                ],
                "tags": [
                    "Templates"
                ],
                "summary": "Adds or updates a template",
                "parameters": [
                    {
                        "description": "Template",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ServerDefinition"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Template name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OAuth2Application": [
                            "templates.repo.delete"
                        ]
                    }
                ],
                "description": "Deletes template",
                "tags": [
                    "Templates"
                ],
                "summary": "Deletes template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/templates/{repo}": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": [
                            "templates.view"
                        ]
                    }
                ],
                "description": "Gets all templates from a repository",
                "tags": [
                    "Templates"
                ],
                "summary": "Get all templates from repo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Repo id",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Template"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OAuth2Application": [
                            "templates.repo.delete"
                        ]
                    }
                ],
                "description": "Deletes a repo from the service",
                "tags": [
                    "Templates"
                ],
                "summary": "Delete repo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Repo Id",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/templates/{repo}/{template}": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": [
                            "templates.view"
                        ]
                    }
                ],
                "description": "Gets a template from the repo",
                "tags": [
                    "Templates"
                ],
                "summary": "Get template",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Repo Id",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Template name",
                        "name": "template",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Template"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/Error"
                }
            }
        },
        "Execution": {
            "type": "object",
            "properties": {
                "autorecover": {
                    "type": "boolean"
                },
                "autorestart": {
                    "type": "boolean"
                },
                "autostart": {
                    "type": "boolean"
                },
                "command": {},
                "environmentVars": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "expectedExitCode": {
                    "type": "integer"
                },
                "post": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MetadataWithIf"
                    }
                },
                "pre": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MetadataWithIf"
                    }
                },
                "stdin": {
                    "$ref": "#/definitions/StdinConsoleConfiguration"
                },
                "stop": {
                    "type": "string"
                },
                "stopCode": {
                    "type": "integer"
                },
                "workingDirectory": {
                    "type": "string"
                }
            }
        },
...
        "Group": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "display": {
                    "type": "string"
                },
                "if": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "variables": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "KeepAlive": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                }
            }
        },
        "Metadata": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                }
            }
        },
        "MetadataWithIf": {
            "type": "object",
            "properties": {
                "if": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "Requirements": {
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string"
                },
                "binaries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "os": {
                    "type": "string"
                }
            }
        },
...
        "ServerDefinition": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/Variable"
                    }
                },
                "display": {
                    "type": "string"
                },
                "environment": {
                    "$ref": "#/definitions/Metadata"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Group"
                    }
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "install": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MetadataWithIf"
                    }
                },
                "keepAlive": {
                    "$ref": "#/definitions/KeepAlive"
                },
                "query": {
                    "$ref": "#/definitions/Metadata"
                },
                "requirements": {
                    "$ref": "#/definitions/Requirements"
                },
                "run": {
                    "$ref": "#/definitions/Execution"
                },
                "stats": {
                    "$ref": "#/definitions/Metadata"
                },
                "supportedEnvironments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Metadata"
                    }
                },
                "type": {
                    "type": "string"
                },
                "uninstall": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MetadataWithIf"
                    }
                }
            }
        },
...
        "StdinConsoleConfiguration": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "Template": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/Variable"
                    }
                },
                "display": {
                    "type": "string"
                },
                "environment": {
                    "$ref": "#/definitions/Metadata"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Group"
                    }
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "install": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MetadataWithIf"
                    }
                },
                "keepAlive": {
                    "$ref": "#/definitions/KeepAlive"
                },
                "name": {
                    "type": "string"
                },
                "query": {
                    "$ref": "#/definitions/Metadata"
                },
                "readme": {
                    "type": "string"
                },
                "requirements": {
                    "$ref": "#/definitions/Requirements"
                },
                "run": {
                    "$ref": "#/definitions/Execution"
                },
                "stats": {
                    "$ref": "#/definitions/Metadata"
                },
                "supportedEnvironments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Metadata"
                    }
                },
                "type": {
                    "type": "string"
                },
                "uninstall": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MetadataWithIf"
                    }
                }
            }
        },
        "TemplateRepo": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isLocal": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "Variable": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "display": {
                    "type": "string"
                },
                "internal": {
                    "type": "boolean"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/VariableOption"
                    }
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "userEdit": {
                    "type": "boolean"
                },
                "value": {}
            }
        },
        "VariableOption": {
            "type": "object",
            "properties": {
                "display": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
    }
}